added engine

+++ src/piece.rs
@@ -1,12 +1,12 @@
-pub static NONE: u8 = 0;
-pub static PAWN: u8 = 1;
-pub static BISHOP: u8 = 2;
-pub static KNIGHT: u8 = 3;
-pub static ROOK: u8 = 4;
-pub static QUEEN: u8 = 5;
-pub static KING: u8 = 6;
+pub static NONE: i32 = 0;
+pub static PAWN: i32 = 1;
+pub static BISHOP: i32 = 2;
+pub static KNIGHT: i32 = 3;
+pub static ROOK: i32 = 4;
+pub static QUEEN: i32 = 5;
+pub static KING: i32 = 6;
 
-pub static WHITE_TEAM: u8 = 8;
-pub static BLACK_TEAM: u8 = 16;
+pub static WHITE_TEAM: i32 = 8;
+pub static BLACK_TEAM: i32 = 16;
 
 // to extract the first 3 bits you can do [num] % [2^(3)]
